Move FakeRust classes and factories to a fixtures package
