+++ libraries/matrix/impl/src/test/kotlin/io/element/android/libraries/matrix/impl/fixtures/RustRoomMember.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2024 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ * Please see LICENSE in the repository root for full details.
+ */
+
+package io.element.android.libraries.matrix.impl.fixtures
+
+import io.element.android.libraries.matrix.api.core.UserId
+import org.matrix.rustcomponents.sdk.MembershipState
+import org.matrix.rustcomponents.sdk.RoomMember
+import uniffi.matrix_sdk.RoomMemberRole
+
+fun aRustRoomMember(
+    userId: UserId,
+    displayName: String? = null,
+    avatarUrl: String? = null,
+    membership: MembershipState = MembershipState.JOIN,
+    isNameAmbiguous: Boolean = false,
+    powerLevel: Long = 0L,
+    isIgnored: Boolean = false,
+    role: RoomMemberRole = RoomMemberRole.USER,
+) = RoomMember(
+    userId = userId.value,
+    displayName = displayName,
+    avatarUrl = avatarUrl,
+    membership = membership,
+    isNameAmbiguous = isNameAmbiguous,
+    powerLevel = powerLevel,
+    normalizedPowerLevel = powerLevel,
+    isIgnored = isIgnored,
+    suggestedRoleForPowerLevel = role,
+)
