+++ libraries/matrix/impl/src/test/kotlin/io/element/android/libraries/matrix/impl/room/member/RoomMemberListFetcherTest.kt
@@ -9,13 +9,14 @@ package io.element.android.libraries.matrix.impl.room.member
 
 import app.cash.turbine.test
 import com.google.common.truth.Truth.assertThat
-import io.element.android.libraries.matrix.api.core.UserId
 import io.element.android.libraries.matrix.api.room.MatrixRoomMembersState
 import io.element.android.libraries.matrix.api.room.roomMembers
+import io.element.android.libraries.matrix.impl.fixtures.FakeRustRoom
+import io.element.android.libraries.matrix.impl.fixtures.FakeRustRoomMembersIterator
+import io.element.android.libraries.matrix.impl.fixtures.aRustRoomMember
 import io.element.android.libraries.matrix.impl.room.member.RoomMemberListFetcher.Source.CACHE
 import io.element.android.libraries.matrix.impl.room.member.RoomMemberListFetcher.Source.CACHE_AND_SERVER
 import io.element.android.libraries.matrix.impl.room.member.RoomMemberListFetcher.Source.SERVER
-import io.element.android.libraries.matrix.test.A_ROOM_ID
 import io.element.android.libraries.matrix.test.A_USER_ID
 import io.element.android.libraries.matrix.test.A_USER_ID_2
 import io.element.android.libraries.matrix.test.A_USER_ID_3
@@ -23,22 +24,16 @@ import io.element.android.libraries.matrix.test.A_USER_ID_4
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
-import org.matrix.rustcomponents.sdk.MembershipState
-import org.matrix.rustcomponents.sdk.NoPointer
-import org.matrix.rustcomponents.sdk.Room
-import org.matrix.rustcomponents.sdk.RoomMember
-import org.matrix.rustcomponents.sdk.RoomMembersIterator
-import uniffi.matrix_sdk.RoomMemberRole
 
 class RoomMemberListFetcherTest {
     @Test
     fun `fetchRoomMembers with CACHE source - emits cached members, if any`() = runTest {
         val room = FakeRustRoom(getMembersNoSync = {
-            FakeRoomMembersIterator(
+            FakeRustRoomMembersIterator(
                 listOf(
-                    fakeRustRoomMember(A_USER_ID),
-                    fakeRustRoomMember(A_USER_ID_2),
-                    fakeRustRoomMember(A_USER_ID_3),
+                    aRustRoomMember(A_USER_ID),
+                    aRustRoomMember(A_USER_ID_2),
+                    aRustRoomMember(A_USER_ID_3),
                 )
             )
         })
@@ -65,7 +60,7 @@ class RoomMemberListFetcherTest {
     @Test
     fun `fetchRoomMembers with CACHE source - emits empty list, if no members exist`() = runTest {
         val room = FakeRustRoom(getMembersNoSync = {
-            FakeRoomMembersIterator(emptyList())
+            FakeRustRoomMembersIterator(emptyList())
         })
 
         val fetcher = RoomMemberListFetcher(room, Dispatchers.Default)
@@ -95,11 +90,11 @@ class RoomMemberListFetcherTest {
     @Test
     fun `fetchRoomMembers with CACHE source - emits all items at once`() = runTest {
         val room = FakeRustRoom(getMembersNoSync = {
-            FakeRoomMembersIterator(
+            FakeRustRoomMembersIterator(
                 listOf(
-                    fakeRustRoomMember(A_USER_ID),
-                    fakeRustRoomMember(A_USER_ID_2),
-                    fakeRustRoomMember(A_USER_ID_3),
+                    aRustRoomMember(A_USER_ID),
+                    aRustRoomMember(A_USER_ID_2),
+                    aRustRoomMember(A_USER_ID_3),
                 )
             )
         })
@@ -122,11 +117,11 @@ class RoomMemberListFetcherTest {
     @Test
     fun `fetchRoomMembers with SERVER source - emits only new members, if any`() = runTest {
         val room = FakeRustRoom(getMembers = {
-            FakeRoomMembersIterator(
+            FakeRustRoomMembersIterator(
                 listOf(
-                    fakeRustRoomMember(A_USER_ID),
-                    fakeRustRoomMember(A_USER_ID_2),
-                    fakeRustRoomMember(A_USER_ID_3),
+                    aRustRoomMember(A_USER_ID),
+                    aRustRoomMember(A_USER_ID_2),
+                    aRustRoomMember(A_USER_ID_3),
                 )
             )
         })
@@ -163,14 +158,14 @@ class RoomMemberListFetcherTest {
     fun `fetchRoomMembers with CACHE_AND_SERVER source - returns cached items first, then new ones`() = runTest {
         val room = FakeRustRoom(
             getMembersNoSync = {
-                FakeRoomMembersIterator(listOf(fakeRustRoomMember(A_USER_ID_4)))
+                FakeRustRoomMembersIterator(listOf(aRustRoomMember(A_USER_ID_4)))
             },
             getMembers = {
-                FakeRoomMembersIterator(
+                FakeRustRoomMembersIterator(
                     listOf(
-                        fakeRustRoomMember(A_USER_ID),
-                        fakeRustRoomMember(A_USER_ID_2),
-                        fakeRustRoomMember(A_USER_ID_3),
+                        aRustRoomMember(A_USER_ID),
+                        aRustRoomMember(A_USER_ID_2),
+                        aRustRoomMember(A_USER_ID_3),
                     )
                 )
             }
@@ -203,69 +198,3 @@ class RoomMemberListFetcherTest {
         }
     }
 }
-
-class FakeRustRoom(
-    private val getMembers: () -> RoomMembersIterator = { FakeRoomMembersIterator() },
-    private val getMembersNoSync: () -> RoomMembersIterator = { FakeRoomMembersIterator() },
-) : Room(NoPointer) {
-    var membersCallCount = 0
-    var membersNoSyncCallCount = 0
-
-    override fun id(): String {
-        return A_ROOM_ID.value
-    }
-
-    override suspend fun members(): RoomMembersIterator {
-        membersCallCount++
-        return getMembers()
-    }
-
-    override suspend fun membersNoSync(): RoomMembersIterator {
-        membersNoSyncCallCount++
-        return getMembersNoSync()
-    }
-
-    override fun close() {
-        // No-op
-    }
-}
-
-class FakeRoomMembersIterator(
-    private var members: List<RoomMember>? = null
-) : RoomMembersIterator(NoPointer) {
-    override fun len(): UInt {
-        return members?.size?.toUInt() ?: 0u
-    }
-
-    override fun nextChunk(chunkSize: UInt): List<RoomMember>? {
-        if (members?.isEmpty() == true) {
-            return null
-        }
-        return members?.let {
-            val result = it.take(chunkSize.toInt())
-            members = it.subList(result.size, it.size)
-            result
-        }
-    }
-}
-
-private fun fakeRustRoomMember(
-    userId: UserId,
-    displayName: String? = null,
-    avatarUrl: String? = null,
-    membership: MembershipState = MembershipState.JOIN,
-    isNameAmbiguous: Boolean = false,
-    powerLevel: Long = 0L,
-    isIgnored: Boolean = false,
-    role: RoomMemberRole = RoomMemberRole.USER,
-) = RoomMember(
-    userId = userId.value,
-    displayName = displayName,
-    avatarUrl = avatarUrl,
-    membership = membership,
-    isNameAmbiguous = isNameAmbiguous,
-    powerLevel = powerLevel,
-    normalizedPowerLevel = powerLevel,
-    isIgnored = isIgnored,
-    suggestedRoleForPowerLevel = role,
-)
