+++ libraries/matrix/impl/src/test/kotlin/io/element/android/libraries/matrix/impl/roomlist/RoomSummaryListProcessorTest.kt
@@ -9,12 +9,10 @@ package io.element.android.libraries.matrix.impl.roomlist
 
 import com.google.common.truth.Truth.assertThat
 import com.sun.jna.Pointer
-import io.element.android.libraries.matrix.api.core.RoomId
-import io.element.android.libraries.matrix.api.core.UserId
 import io.element.android.libraries.matrix.api.roomlist.RoomSummary
+import io.element.android.libraries.matrix.impl.fixtures.FakeRustRoomListItem
 import io.element.android.libraries.matrix.test.A_ROOM_ID
 import io.element.android.libraries.matrix.test.A_ROOM_ID_2
-import io.element.android.libraries.matrix.test.A_ROOM_NAME
 import io.element.android.libraries.matrix.test.room.aRoomSummary
 import io.element.android.libraries.matrix.test.room.aRoomSummaryFilled
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -22,19 +20,12 @@ import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
-import org.matrix.rustcomponents.sdk.EventTimelineItem
-import org.matrix.rustcomponents.sdk.Membership
-import org.matrix.rustcomponents.sdk.NoPointer
-import org.matrix.rustcomponents.sdk.RoomHero
-import org.matrix.rustcomponents.sdk.RoomInfo
 import org.matrix.rustcomponents.sdk.RoomList
 import org.matrix.rustcomponents.sdk.RoomListEntriesUpdate
 import org.matrix.rustcomponents.sdk.RoomListItem
 import org.matrix.rustcomponents.sdk.RoomListServiceInterface
 import org.matrix.rustcomponents.sdk.RoomListServiceStateListener
 import org.matrix.rustcomponents.sdk.RoomListServiceSyncIndicatorListener
-import org.matrix.rustcomponents.sdk.RoomMember
-import org.matrix.rustcomponents.sdk.RoomNotificationMode
 import org.matrix.rustcomponents.sdk.RoomSubscription
 import org.matrix.rustcomponents.sdk.TaskHandle
 
@@ -48,7 +39,7 @@ class RoomSummaryListProcessorTest {
         summaries.value = listOf(aRoomSummary())
         val processor = createProcessor()
 
-        val newEntry = FakeRoomListItem(A_ROOM_ID_2)
+        val newEntry = FakeRustRoomListItem(A_ROOM_ID_2)
         processor.postUpdate(listOf(RoomListEntriesUpdate.Append(listOf(newEntry, newEntry, newEntry))))
 
         assertThat(summaries.value.count()).isEqualTo(4)
@@ -59,7 +50,7 @@ class RoomSummaryListProcessorTest {
     fun `PushBack adds a new entry at the end of the list`() = runTest {
         summaries.value = listOf(aRoomSummaryFilled())
         val processor = createProcessor()
-        processor.postUpdate(listOf(RoomListEntriesUpdate.PushBack(FakeRoomListItem(A_ROOM_ID_2))))
+        processor.postUpdate(listOf(RoomListEntriesUpdate.PushBack(FakeRustRoomListItem(A_ROOM_ID_2))))
 
         assertThat(summaries.value.count()).isEqualTo(2)
         assertThat(summaries.value.last().roomId).isEqualTo(A_ROOM_ID_2)
@@ -69,7 +60,7 @@ class RoomSummaryListProcessorTest {
     fun `PushFront inserts a new entry at the start of the list`() = runTest {
         summaries.value = listOf(aRoomSummaryFilled())
         val processor = createProcessor()
-        processor.postUpdate(listOf(RoomListEntriesUpdate.PushFront(FakeRoomListItem(A_ROOM_ID_2))))
+        processor.postUpdate(listOf(RoomListEntriesUpdate.PushFront(FakeRustRoomListItem(A_ROOM_ID_2))))
 
         assertThat(summaries.value.count()).isEqualTo(2)
         assertThat(summaries.value.first().roomId).isEqualTo(A_ROOM_ID_2)
@@ -81,7 +72,7 @@ class RoomSummaryListProcessorTest {
         val processor = createProcessor()
         val index = 0
 
-        processor.postUpdate(listOf(RoomListEntriesUpdate.Set(index.toUInt(), FakeRoomListItem(A_ROOM_ID_2))))
+        processor.postUpdate(listOf(RoomListEntriesUpdate.Set(index.toUInt(), FakeRustRoomListItem(A_ROOM_ID_2))))
 
         assertThat(summaries.value.count()).isEqualTo(1)
         assertThat(summaries.value[index].roomId).isEqualTo(A_ROOM_ID_2)
@@ -93,7 +84,7 @@ class RoomSummaryListProcessorTest {
         val processor = createProcessor()
         val index = 0
 
-        processor.postUpdate(listOf(RoomListEntriesUpdate.Insert(index.toUInt(), FakeRoomListItem(A_ROOM_ID_2))))
+        processor.postUpdate(listOf(RoomListEntriesUpdate.Insert(index.toUInt(), FakeRustRoomListItem(A_ROOM_ID_2))))
 
         assertThat(summaries.value.count()).isEqualTo(2)
         assertThat(summaries.value[index].roomId).isEqualTo(A_ROOM_ID_2)
@@ -183,87 +174,5 @@ class RoomSummaryListProcessorTest {
         }
 
         override fun subscribeToRooms(roomIds: List<String>, settings: RoomSubscription?) = Unit
-    }
-}
-
-private fun aRustRoomInfo(
-    id: String = A_ROOM_ID.value,
-    displayName: String = A_ROOM_NAME,
-    rawName: String = A_ROOM_NAME,
-    topic: String? = null,
-    avatarUrl: String? = null,
-    isDirect: Boolean = false,
-    isPublic: Boolean = false,
-    isSpace: Boolean = false,
-    isTombstoned: Boolean = false,
-    isFavourite: Boolean = false,
-    canonicalAlias: String? = null,
-    alternativeAliases: List<String> = listOf(),
-    membership: Membership = Membership.JOINED,
-    inviter: RoomMember? = null,
-    heroes: List<RoomHero> = listOf(),
-    activeMembersCount: ULong = 0uL,
-    invitedMembersCount: ULong = 0uL,
-    joinedMembersCount: ULong = 0uL,
-    userPowerLevels: Map<String, Long> = mapOf(),
-    highlightCount: ULong = 0uL,
-    notificationCount: ULong = 0uL,
-    userDefinedNotificationMode: RoomNotificationMode? = null,
-    hasRoomCall: Boolean = false,
-    activeRoomCallParticipants: List<String> = listOf(),
-    isMarkedUnread: Boolean = false,
-    numUnreadMessages: ULong = 0uL,
-    numUnreadNotifications: ULong = 0uL,
-    numUnreadMentions: ULong = 0uL,
-    pinnedEventIds: List<String> = listOf(),
-    roomCreator: UserId? = null,
-) = RoomInfo(
-    id = id,
-    displayName = displayName,
-    rawName = rawName,
-    topic = topic,
-    avatarUrl = avatarUrl,
-    isDirect = isDirect,
-    isPublic = isPublic,
-    isSpace = isSpace,
-    isTombstoned = isTombstoned,
-    isFavourite = isFavourite,
-    canonicalAlias = canonicalAlias,
-    alternativeAliases = alternativeAliases,
-    membership = membership,
-    inviter = inviter,
-    heroes = heroes,
-    activeMembersCount = activeMembersCount,
-    invitedMembersCount = invitedMembersCount,
-    joinedMembersCount = joinedMembersCount,
-    userPowerLevels = userPowerLevels,
-    highlightCount = highlightCount,
-    notificationCount = notificationCount,
-    cachedUserDefinedNotificationMode = userDefinedNotificationMode,
-    hasRoomCall = hasRoomCall,
-    activeRoomCallParticipants = activeRoomCallParticipants,
-    isMarkedUnread = isMarkedUnread,
-    numUnreadMessages = numUnreadMessages,
-    numUnreadNotifications = numUnreadNotifications,
-    numUnreadMentions = numUnreadMentions,
-    pinnedEventIds = pinnedEventIds,
-    creator = roomCreator?.value,
-)
-
-class FakeRoomListItem(
-    private val roomId: RoomId,
-    private val roomInfo: RoomInfo = aRustRoomInfo(id = roomId.value),
-    private val latestEvent: EventTimelineItem? = null,
-) : RoomListItem(NoPointer) {
-    override fun id(): String {
-        return roomId.value
-    }
-
-    override suspend fun roomInfo(): RoomInfo {
-        return roomInfo
-    }
-
-    override suspend fun latestEvent(): EventTimelineItem? {
-        return latestEvent
     }
 }
